+++ NEWS.md
@@ -1,3 +1,5 @@
+# nestedmodels (development version)
+
 # nestedmodels 1.0.4
 
 * Fixed the prep.step_nest() method to include an ellipsis argument, to ensure consistency with the generic.